dataplex_client.py
from google.cloud import dataplex_v1

class DataplexClient:
    def __init__(self, project_id, location):
        self.client = dataplex_v1.DataQualityServiceClient()
        self.parent = f"projects/{project_id}/locations/{location}"

    def create_or_update_dq_scan(self, scan_id, table_path, rules):
        # Qui verrebbe implementata la costruzione dell'oggetto DataQualityScan
        # che invia le regole row_count_expectation definite nel YAML
        print(f"[API] Configurazione scansione {scan_id} per {table_path} inviata.")