requirements.txt
google-cloud-dataplex
pyyaml